bluetooth: Add permissions to desktop Page Info
This change enables the desktop Page Info dialog box to display
Bluetooth device permissions if the WebBluetoothNewPermissionsBackend
feature is enabled.  Device permissions can be revoked from this dialog
box.  Additionally, if the Bluetooth guard permission is blocked by the
user (once that capability is enabled with the inclusion of this
permission in Site Settings) or by an enterprise policy, it will be
reflected in the Page Info dialog box.

Design doc:
https://docs.google.com/document/d/1vlTlsQKThwaX0-lj_iZbVTzyqY7LioqERU8DK3u3XjI/edit#bookmark=kix.p5nalpch13qw

Bug: 689240, 570250
Change-Id: I4d972ef051afc0551d99b00d10e229779c01662a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2042259
Commit-Queue: Ovidio de Jesús Ruiz-Henríquez <odejesush@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739657}