Enable WCG on Android under flag
Previously, the flag to enable WCG on Android was inside
compositor_impl_android.cc. But this is too late in the pipeline.
The display color gamuts are already set to wide or not before this.

This CL moves the flag forward so the display color spaces are set
correctly based on what the display actually supports.

Bug: 955158
Change-Id: I5e81b4d711eced61182f1a0466ddb65b00219af5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2036917
Reviewed-by: Khushal <khushalsagar@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: ccameron <ccameron@chromium.org>
Commit-Queue: Chris Blume <cblume@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739677}