CachedMetadataHandler: rename confusing kSendToPlatform
Rename kSendToPlatform to kClearPersistentStorage. The notion of
'platform' seems to be disappearing from Blink, so 'sending to platform'
becomes more and more confusing.

The new name emphasizes that the constant only applies to _clearing_ the
storage, consistently with the remaining constants in the enumeration.
We recently made the 'Set' operation to always persist the changes,
hence we only have an option to persist in the 'Clear' operation. This
way 'Clear' in the name is clearer than 'Send' (pun intended).

Bug: 1045052
Change-Id: I315a3a01c3ac3f97184594674ea260f2508d0b8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2042096
Commit-Queue: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739680}