Complete frames in OpenXR when WebXR submits missing frames.
There are various scenarios where WebXR submits frames marked as
'missing', such as when the visibility state becomes hidden or if
Javascript does not draw to the WebGL layer. However, OpenXR has
already acquired a swapchain and started a frame. When
XRCompositorCommon processes this missing frame, it does not notify
OpenXR to release the swapchain and end the frame. There are a limited
number of swapchains, so after a few repetitions of these scenarios,
OpenXR can no longer acquire a swapchain and render anything,
resulting in a black screen with a loading logo.

With this fix, XRCompositorCommon notifies OpenXrApiWrapper to end the
frame if a missing frame is submitted.

Bug: 1049873
Change-Id: Ic0b88cf952717d87c606f968ea4df67289acafdd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2040908
Commit-Queue: Patrick To <patrto@microsoft.com>
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739689}