UMA for passive detection of breakage caused by "allowlist for cors".
This CL adds detection of breakage that may be caused by enabling CORS
for content scripts of Chrome extensions - this is controlled by the
CorbAllowlistAlsoAppliesToOorCors feature.  The detection works in
"passive" mode - without actually modifying the CORS behavior in Chrome.

The "passive" detection reports that there is a risk that a response
might be broken if all of the 3 conditions below are true:
1. The request is triggered by a content script of a Chrome Extension
2. The response is allowed by CORB (we expect that extensions triggering
   responses blocked by CORB are already allowlisted and therefore won't
   be affected by the planned CORS changes)
3. The response would be blocked by CORS if it was enabled (i.e. the
   request was cross-origin, in `cors` mode and the response didn't have
   a valid `Access-Control-Allow-Origin` response header).
The "passive" mode is overly pessimistic - the responses might actually
be allowed if Chrome's behavior was modified to sent `Origin` request
header which might prompt the HTTP server to reply with a proper
`Access-Control-Allow-Origin` response header (see the 3rd item above).

The CL consists of 3 parts:
- Moving the definition of the CorbAllowlistAlsoAppliesToOorCors feature
  to the //services/network layer.
- Declaration and implementation of the new UMA
- Test coverage

Bug: 920638
Change-Id: I6656d70925f63235c91430441dff978e4d747c94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2033639
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739693}