grit: util: use io.open to read files w/encoding
The trick of opening a file, reading it, and then decoding it doesn't
work in Python 3.  If it isn't read as a binary stream, then it will
be decoded immediately, and if no encoding is specified at open time,
then it will use the env/system default (which could be anything).

Lets switch to using the io.open helper which allows us to pass the
encoding setting directly in Python 2 & Python 3.  This also allows
us to simplify the code a bit which is nice.

Bug: 983071
Test: `./grit/test_suite_all.py` passes
Change-Id: I762b1c34ea6937a90377c26863537be3654c76b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2043314
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739708}