[cascade] Add CascadeOrigin and CascadePriority
In preparation of the StyleCascade rework, CascadeOrigin and
CascadePriority are added without being used yet (except by tests).

CascadeOrigin is the new version of StyleCascade::Origin. Unlike
CascadeOrigin, it does not encapsulate 'important', because importance
is not an aspect of a rule/declaration-block, but rather an aspect of
a declaration. This would make it awkward to store StyleCascade::Origin
on the rule-level (which is coming in a later CL), since there are
values of StyleCascade::Origin that don't make sense on that level (e.g.
kUserAgentImportant).

Importance is now instead handled directly by CascadePriority (the new
StyleCascade::Priority). CascadePriority encodes all the cascade
criteria into a single uint64_t, such that two priorities can be checked
against each other by comparing the integer.

CascadePriority is also different in that it now also includes a
position component. The position is an "address" to the declaration in
the MatchResult. It provides a way to randomly access a declaration's
CSSValue "in-place" without actually storing the CSSValue in the
cascade.

BUG=947004

Change-Id: I31429ce86576b29a68968298e6a47870b0baad1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2041612
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739761}