[quickview] Close Quick View on check-select change
Currently when you delete all of the items in a check-selection from
Quick View, but there are still files in the file list, Quick View will
not close but instead show the next file in the list.

Fix this: Make Quick View close when all items in a check selection have
been deleted.

Test: Test coming in a future patch.
Bug: 803259
Change-Id: I73a50898e94cc13605e91e365643a2c3b82cbccd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2043515
Commit-Queue: Noel Gordon <noel@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739808}