[Web Payment] Friendlier error message.
Before this patch, omitting "serviceworker" section in the web app
manifest would result in a message "'serviceworker' must be a
dictionary", which is too harsh.

This patch adds "in your web app manifest" to the end of the error
message about the serviceworker section missing from the web app
manifest.

After this patch, omitting "serviceworker" section in the web app
manifest results in a message "'serviceworker' must be a dictionary in
your web app manifest."

Bug: 1050309
Change-Id: I91f49a0daadf65509a4a756fe597792ca86aa25c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2045478
Reviewed-by: Liquan (Max) Gu <maxlg@chromium.org>
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739845}