Re-enable InstallNewExtension test on Mac
Re-enable the ComponentCloudPolicyTest.InstallNewExtension browser test
on macOS.

The test was disabled due to flakiness a long time ago, but no logs
survived from that failure and also the test passed the local
stress-test with ~1'000 iterations on Mac (and 10'000 iterations on
Linux). It means that, hopefully, the flakiness has been fixed
elsewhere.

Bug: 816647
Change-Id: I127b19846959207734142d5c80943883fa4215cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2044099
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739848}