Ensure the quiet notification prompt IPH is only displayed for...
... notification mini-infobar prompts.

Follow up to:
https://chromium-review.googlesource.com/c/chromium/src/+/1944173

While currently notifications are the only permission that might use
a mini-infobar as the prompt, this will very likely change. Because
the logic that dictates when this IPH is shown is easy to miss
it's likely new permissions will still have the IPH. The solution is
to ensure that only notification permission mini-infobars provide
IPH tracker parameters.

This CL adds an extra check to ensure that the current permission
request is a notification permission request when checking if the
IPH should be shown.

Bug: 1029595
Change-Id: Iceca6765f3e9764e802af69c4f8cddd705eedf03
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1948744
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Andy Paicu <andypaicu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739863}