Don't record restricted ad tagging histogram for frames with zero bytes
The not ignored bucket of this histogram should equal
PageLoad.Clients.Ads.FrameCounts.AdFrames.Total. It currently does not
because this is called prior to checking ShouldRecordFrameForMetrics().

This change does affect how the metric is recorded, but the metrics was
newly added so versioning is not as ncesssary.

Bug: 1030325
Change-Id: Ia07b24847dffab30af06161b2d17e63c532f73a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2042032
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Commit-Queue: John Delaney <johnidel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739906}