[cascade] Non-persistent StyleCascade
The first attempt at implementing StyleCascade used a HashMap to
store CSSPropertyName-to-Value entries for both standard and custom
properties. The CSSValues were added directly into this HashMap, and
during Apply we would simply traverse the HashMap backing and apply
them. This resulted in pretty nice and clear API. Unfortunately it also
resulted in severe performance regressions due to HashMap overhead.

This CL tries another approach, which should have performance
characteristics similar to that of master: instead iterating through
the MatchResult and _adding_ values (CSSValues) to the cascade,
the MatchResult is simply _analyzed_. This means that we don't add
any CSSValue to the cascade, but simply "cascade information"
(a CascadePriority object) about each property seen. This object
contains information about where the CSSValue can be found in the
MatchResult, such that we can still Apply declarations with random
access, if needed.

Since no CSSValues are stored in the StyleCacade, it can't Apply the
values on its own. Instead, we use the previous analysis of the
MatchResult on our second pass: this time we know whether a given
declaration should be applied or skipped (for example, we know whether
a declaration with higher priority will appear later).

Declarations are also no longer stored in a HashMap, but in a
CascadeMap, which stores standard properties in a fixed-size array. The
hope is that this, along with traversal of the MatchResult rather than
the traversal of the hash backing, will have similar performance to that
of master.

Notes, in no particular order:

 - Instead of traversing the MatchResult and adding CSSValues to the
   StyleCascade, the MatchResult is now _analyzed_, which means that
   figure out which declaration should win the cascade for a given
   property, simply store the position of that declaration in the
   MatchResult.
 - Interpolations work differently too. Instead of calling out to
   an Animator, we traverse the active interpolations similarly to how
   it takes place on master currently. We use a prior call to Analyze to
   figure out which interpolations should actually apply, and which
   should be skipped.
 - CascadeInterpolations is a simple transport-structure for
   ActiveInterpolationMaps. We need to provide anything from 0 to 4
   ActiveInterpolationMaps, depending on the circumstance. These are
   put in a vector, such that CascadePriority can encode a meaningful
   position for the interpolations. You can think of
   CascadeInterpolations as a MatchResult, except for interpolations.
 - The Exclude/filter mechanism has been reworked. A CascadeFilter must
   be explicitly passed to Analyze/Apply. Previously, exclusions only
   had an affect on the Apply pass, but now the caller can choose to
   filter either during Analyze, or Apply, or both.
 - There's a new optimization regarding kHighPropertyPriority: the
   CascadeMap keeps track of seen properties kHighPropertyPriority in a
   uint64_t. This can be used to figure out which high-prio properties
   have been seen if (any), allowing us to avoid calls to Apply for
   properties without the corresponding bit set.
 - Apply functions that lookup a declaration with random access have
   been renamed to LookupAndApply to make this more clear.
 - The CSSPendingSubstitution resolving has changed. Previously, there
   was a mechanism which allowed this code to re-use the result of the
   shorthand parsing on all the relevant longhands with the same
   CSSPendingSubstitution present in the cascade. However, since we no
   longer store the actual value in the cascade, this is impractical.
   Now we take a simpler approach: the most recently parsed shorthand
   is cached on Resolver, and if the next call to ResolvePending-
   Substitution needs to parse the same value, we can just fetch it from
   the cache.
 - There was a bug related to
   SetHasVariableReferenceFromNonInheritedProperty. I was attempting to
   marked the referenced property, when I should have marked the
   property that contained the reference. This was discovered because
   this CL always re-introduces inherited-only cache hits (which turned)
   out to be a significant performance benefit (unfortunately).

Bug: 947004
Change-Id: I2808e407aa3a5d2cad8867a7f03a7261adca006a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2041656
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739924}