[Extensions Menu] Reland hide extensions container when none are installed
Original CL courtesy of pbos@ in revision 9b92a6a215.

Original description:
This change makes ExtensionsToolbarContainer invisible by default. It
turns visible when any extension is added and goes invisible if the last
extension is removed.

Adjustments:
The original CL was reverted because it made the Extensions Menu
browsertests flaky. This is because, when the container was
transitioning between invisible and visible, we didn't wait for
the layout to complete. This resulted in the extensions menu trying
to anchor to the wrong point, and the tests failing.

Fix this by force a layout when adding or removing extensions.

Bug: 943702

Change-Id: I791b3762543e53222219a52eba67ddb40bcf947e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2044336
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Peter Boström <pbos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739933}