cros: Update TurnSyncOnHelper to work with unconsented primary account
TurnSyncOnHelper is responsible for showing the sync consent dialog
when the first browser window is opened. Now that Chrome OS has been
converted so it doesn't set a sync-consented primary account on login
we can use GetPrimaryAccountInfo() to determine if the user has
consented to sync or not.

Update TurnSyncOnHelper to check IdentityManager for sync consent.
This requires updating SyncConfirmationHandler to work with the
unconsented primary account (otherwise TurnSyncOnHelper would have
to set a consented primary account when the dialog opens, then
revoke consent when the dialog closes).

If the user consents to sync, upgrade their unconsented primary
account to a consented primary account.

Bug: 1042400, 1013466
Test: updated unit_tests
Change-Id: I94579b88a2e986c6157069788c3b53938608ef1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2044065
Commit-Queue: James Cook <jamescook@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739963}