[Switch Access] Support opening existing tabs
Tabs are actionable (click to open), but they also contain a button
(to close the tab). This is a structure not really found elsewhere in
the accessibility tree, so this change adds a custom node type TabNode
to handle the logic of showing the tab group and tab actions.

Bug: 1022987
Change-Id: Iec8914d2b847100b6e4fa2fbd6832bbee529fcd3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1907312
Commit-Queue: Anastasia Helfinstein <anastasi@google.com>
Reviewed-by: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739967}