[Switch Access] Fix physical key event detection when running ChromeOS on Linux
Switch Access needs to distinguish between physical and virtual key
events. On ChromeOS devices, this is done using the device_id. However,
that field is not populated when running ChromeOS on Linux. The code
previously used the presence of a property KeyboardIBusFlag to identify
events originating from a physical keyboard. However, that property no
longer appears to be set. This CL changes to use a property
KeyboardHwKeyCode which is set.

TBR=dtseng@chromium.org

Bug: None
Change-Id: I2f7c32d7b2679e31f21afa7b3d1c7f08be8e7f7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2047372
Reviewed-by: Anastasia Helfinstein <anastasi@google.com>
Commit-Queue: Anastasia Helfinstein <anastasi@google.com>
Cr-Commit-Position: refs/heads/master@{#739970}