inkdrop: Fix regression where opacity of highlight was always 1.f.
This was caused by a tweak in the default args to accomodate the solid
color layer highlight used by cros (for now). However, one of the
constructors calls another which overrides the default opacity to 1.f,
so we have to manually set it back to .128f after the other constructor
is called.

Test: manual
Change-Id: I4f306f13c2da9e5c09f9d7c65485a6c7cda8016c
Fixed: 1050298
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2047367
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Peter Boström <pbos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739976}