aw: Fix invalidate heuristics for viz
We rely on existing surface damage to decide whether invalidate or not
for subclients. Previously boolean flag was which was problematic
because kModeDraw can happen before or after the next begin frame.
This CL switches to using counter of damages and draws to address this
issue.

Bug: 805739, 1050785
Change-Id: Ibe1f89c4b031178295e57fada43cb032b865ce54
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2047825
Commit-Queue: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740038}