Removes AddRuleForHostname() from ProxyBypassRules
This API is allowing embedders of ProxyBypassRules to add rules directly
without ProxyBypassRules' parsing function, which isn't ideal. We should
have a central function to parse the input from outside of
ProxyBypassRules.

This CL removes AddRuleForHostname() and replace it with
AddRuleFromString(). In issue 1049876, we passed an IP address like
hostname to the internal class SchemeHostPortHostnamePatternRule, which
is now invalid.

Bug: 1049876
Change-Id: I2136d6b03e5cf53a246b7ffd582c158a25e5a8b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2045311
Reviewed-by: Zhongyi Shi <zhongyi@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: David Jacobo <djacobo@chromium.org>
Commit-Queue: Shimi Zhang <ctzsm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740039}