[fuchsia] Fix vulkan with in process GPU thread
The scenic port allows surfaces to be created on any thread, but
some DCHECKs are firing in this case. Just remove the asserts as
gpu_host_ should be accessed on the main thread only; this fails
in tests when there's an in process GPU thread.

Bug: 1047032

Change-Id: Ib47528fb8ded0f2b82d795566f55e950e2d19c8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2044760
Commit-Queue: Michael Spang <spang@chromium.org>
Reviewed-by: Sergey Ulanov <sergeyu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740064}