Fix crostini <-> task manager integration
Since crrev.com/c/1949703 a new flag was added to the vm launch which
caused our logic for determining which process was crostini to fail.

This changes the logic so that we once again correctly find the crostini
process and report it in the task manager.

See also crrev.com/c/2040592 for similar PluginVM bug.

Bug: 1047589
Change-Id: I28e6bfcc02f69d837df34e306a612fc29efb1547
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2041377
Reviewed-by: David Munro <davidmunro@google.com>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Commit-Queue: Nic Hollingum <hollingum@google.com>
Cr-Commit-Position: refs/heads/master@{#740100}