Files app: FilesNg file list header sort icon ripple
Change the file list header sort icon to be <cr-icon-button> instead of
background image mask to have the ripple effect.

To be able to align icon and label wrapped it in a flex container. This
also benefit the display of column header when the columns are very
narrow. This change applies to previous design too, which I have
manually checked visually and functionality.

Move the logic for adding CSS class for sorting from table_header.js to
file_table.js, so all the logic for sorting UI elements are together and
these CSS classes aren't added for the new design so they don't have to
be overwritten in CSS.

Change the minimum width for columns from 10px to 24px, so the dragger
icons don't overlap each other.

Note: The icons are copied from
//ui/file_manager/file_manager/foreground/images/common/arrow_{up|down}_small.svg

Bug: 992821
Change-Id: I17b452585b7639e788d36729fd59b4569f4d928f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2035485
Commit-Queue: Luciano Pacheco <lucmult@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740110}