Guard against URL loader being null?
Given that this signature just stops at OnURLLoadComplete and not into
the callback, I'm pretty sure that the crash has nothing to do with
adding more loaders in the page navigation context. So I'm putting this
null check here just to make sure it doesn't crash. I'm also quite
certain this is in a shutdown flow based on the first thing in the
crash signature so is probably not an actually damaging thing.

Bug: 1050481
Change-Id: I62d0ad06b5decf611cb307a515c81434082bd976
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2047493
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: Michael Crouse <mcrouse@chromium.org>
Commit-Queue: Sophie Chang <sophiechang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740123}