Add placeholder cookies page in redesigned privacy settings
Empty page with banner on which to build the new cookie page in go/friendly-settings-m81.
Tests will be added in future cls that add content and functionality to this
blank page. The page is only navigable directly from the url bar at
chrome://settings/cookies when the privacySettingsRedesignEnabled feature flag
is on.


Bug: 1032584
Change-Id: I66f952ad50a6beffd223160c22a4e47772f30970
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1986828
Commit-Queue: Sean Harrison <harrisonsean@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740128}