[iOS] Don't call HandlerForProtocol before commands are handled.
HandlerForProtocol() CHECKs if the commands are not yet handled by
the command dispatcher.  The previous implementation called this too
early, as the interaction handlers are created upon Browser::Create(),
but the application commands are not yet handled at this point.  This
CL updates the approach to refrain from fetching the command handler
until interaction with the Browser actually occurs.

Bug: 1030357
Change-Id: I803bb6dc844c735c5d9d0dffb798e57ae67f0cec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2031756
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740134}