[iOS] Create OverlayRequestQueueImpl::Delegate
When OverlayRequests are removed from a queue, ownership of the removed
request is transferred to the delegate.  This is not possible using the
observer interface, since there can be multiple observers, and passing
ownership of the unique_ptr to any one of these observers would
invalidate the parameter passed to subsequent observers.

If a request for currently-presented overlay UI is removed from its
queue, OverlayPresenterImpl keeps the removed request in memory until
its dismissal is completed so that the request stays alive until all
observers can be notified that the requested overlay UI is finished
being hidden.

This CL also removes OverlayRequestQueueImpl::PopBackRequest(), since
it is no longer used after this CL.

Bug: none
Change-Id: I6bab3fb1877ea71bf6a3e66f3f7930fa8700b5b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2039593
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740136}