Uniform the scroll offset to be float
In scrollable shelf view code, types of the scroll offset are mixed:
some are integer while some are float. This CL makes sure that the
scroll offset should be float. In some functions like AdjustOffset(),
the integer input is expected since modulo operation is used there. This
CL also ensures to call these functions when it is meaningful to do so.

No UI behavior should be modified,

Bug: 1043275
Change-Id: I7421d8573348d7a0f57c9e8aac38a218f95c53e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2047412
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740137}