Crostini Port Forwarding: Preference Tracking (add/remove)
This CL covers the tracking of port forwarding settings in the user's
profile preferences. The port rule settings are all contained in a flat
list where each rule is a dictionary object like so:
{
	port_number: 5000,
 	protocol_type: 0, (TCP)
	input_ifname: "all",
	label: "some-label",
	active: true,
	container_name: "penguin",
	vm_name: "termina",
}

The preferences are only modified by add / remove operations currently
and activate / deactivate will come in a later CL.

Involves changes in the completion of callbacks as well as lots of new
unit tests. Additionally, the new profile preference is registered in
crostini_pref_names*.

Bug: 848127
Test: crostini_port_forward_unittest.cc (+ASAN)
Change-Id: I65df2f5a9297fc45051adc8c219ede1318861eb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2037300
Reviewed-by: Fergus Dall <sidereal@google.com>
Reviewed-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: David Munro <davidmunro@google.com>
Commit-Queue: Matthew Chen <matterchen@google.com>
Cr-Commit-Position: refs/heads/master@{#740161}