Added a mojo generated Pagehandler for chrome://media-app.
This CL defines a mojom file for the mojo interface
between the Browser and the chrome://media-app page and
implements the relevant CPP (Page Handler) and JS
(mojo_api_bootstrap.js) interfaces. This interface is leveraged to
expose the native browser open feedback dialog functionality
to the page via a MediaAppUIDelegate.

This Cl also adds the kFeedbackSourceMediaApp constant to
the chrome::FeedbackSource enum to allow feedback dialogs
opened by the media app to be correctly logged.

Bug: 996088,1045222
Change-Id: Ia6f28c2330cfb592f9440778b94ebf45da89dbc2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2015874
Commit-Queue: Zain Afzal <zafzal@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Trent Apted <tapted@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740168}